pyproject.toml
[project]
name = "resume-parse-pro-ai"
version = "0.1.0"
description = "Resume Parse Pro AI - AI-powered resume parsing application"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "pdfminer.six>=20221105",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "django-cors-headers>=4.3.0",
    "tenacity>=8.2.0",
    "django-filter>=24.0",
    "docx2txt>=0.8",
]

[project.optional-dependencies]
dev = []

[tool.uv]
dev-dependencies = []

